+++ src/main/java/br/com/compass/pb/libraryual/service/GenreService.java
@@ -2,6 +2,7 @@ package br.com.compass.pb.libraryual.service;
 
 import br.com.compass.pb.libraryual.domain.dto.GenreDTO;
 import br.com.compass.pb.libraryual.domain.entity.Genre;
+import br.com.compass.pb.libraryual.exception.ResourceNotFoundException;
 import br.com.compass.pb.libraryual.repository.GenreRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,14 +15,19 @@ import java.util.stream.Collectors;
 @Service
 public class GenreService {
 
+    private static final String GENRE = "Genre";
+    private static final String GENRE_NOT_FOUND = "Genre not found with ID: ";
     @Autowired
     private GenreRepository genreRepository;
 
     public List<GenreDTO> findAll(){
         List<Genre> genres = genreRepository.findAll();
-        return genres.stream()
-                .map(GenreDTO::convertToDto)
-                .collect(Collectors.toList());
+        if(!genres.isEmpty()) {
+            return genres.stream()
+                    .map(GenreDTO::convertToDto)
+                    .collect(Collectors.toList());
+        }
+        throw new ResourceNotFoundException(GENRE, "There are no records to display");
     }
 
     public GenreDTO findById(Long id) {
@@ -30,7 +36,7 @@ public class GenreService {
             Genre genres = genreOptional.get();
             return GenreDTO.convertToDto(genres);
         }
-        return null;
+        throw new ResourceNotFoundException(GENRE, GENRE_NOT_FOUND + id);
     }
 
     public GenreDTO insert(GenreDTO genreDTO){
@@ -49,10 +55,15 @@ public class GenreService {
             Genre updatedGenre = genreRepository.saveAndFlush(genre);
             return new GenreDTO(updatedGenre);
         }
-        return null;
+        throw new ResourceNotFoundException(GENRE, GENRE_NOT_FOUND + id);
     }
 
     public void delete(Long id){
-        genreRepository.deleteById(id);
+        Optional<Genre> optionalGenre = genreRepository.findById(id);
+        if (optionalGenre.isPresent()) {
+            genreRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException(GENRE, GENRE_NOT_FOUND + id);
+        }
     }
 }
